services/vn_search: Fix incosistency download provider index
When 2 providers are the same but separated but separated by
a different provider like this ``Drive - Terabox - Drive``,
it will shown like this ``Drive 1 Terabox 1 Drive 1`` instead
of ``Drive 1 Terabox 1 Drive 2``.

To fix it, create a Object variable that store the set of
providers used along with their index, and when the
provider is already in that object, increament that
with 1.

Signed-off-by: Taufiq Pohan <m.taufiq30s@hotmail.com>